requirements.txt
djangorestframework-jwt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-filter==2.1.0
django-guardian==1.5.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
docopt==0.6.2
get==2019.3.22
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2
openapi-codec==1.3.2
pipreqs==0.4.9
post==2019.3.22
public==2019.3.22
pytz==2018.9
query-string==2019.3.22
requests==2.21.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9
uritemplate==3.0.0
urllib3==1.24.1
yarg==0.1.9
